+++ packages/plugin-commons/web/src/index.ts
@@ -36,8 +36,7 @@ export { default as RadioGroupHorizontal } from './Components/RadioGroupHorizont
 export { default as RadioGroupVertical } from './Components/RadioGroupVertical';
 export { default as SelectionList } from './Components/SelectionList';
 export { default as SettingsPanelFooter } from './Components/SettingsPanelFooter';
-export { default as Slider } from './Components/Slider';
-export { default as SliderWithInput } from './Components/SliderWithInput';
+export { default as SettingsSection } from './Components/SettingsSection';
 export { Tab, Tabs } from './Components/Tabs';
 export { default as ColorPicker } from './Components/ColorPicker/ColorPicker';